+++ history.md
@@ -6,7 +6,8 @@
   - *added:* Anonymous type support
   - *fixed:* Malformed assmebly names (in certain circumstances)
   - *fixed:* Badly identified declaring types (in certain circumstances)
-- 1.1.2:
+- 2.0.0:
+  - *breaking:* dropped `netstandard1.6` support  
   - *fixed:* Misinterpreted "ref" modifier (in certain circumstances)
   - *fixed:* Proxy generation failed on interface methods having parameter names that were used internally
   - *fixed:* Proxy generation failed on interface methods having by ref array parameter